requirements.txt
pandas
Flask
numpy
scikit-learn
seaborn
ipykernel
pyyaml 
dill
#-e .